# storegw: Optimized inject label stage of index lookup.