--- a/pkg/store/bucket_test.go
+++ b/pkg/store/bucket_test.go
@@ -1257,6 +1257,7 @@ func benchBucketSeries(t testutil.TB, skipChunk bool, samplesPerSeries, totalSer
 			partitioner: gapBasedPartitioner{maxGapSize: partitionerMaxGapSize},
 			chunkObjs:   []string{filepath.Join(id.String(), "chunks", "000001")},
 			chunkPool:   chunkPool,
+			extLset:     extLset,
 		}
 		blocks = append(blocks, b)
 	}
@@ -2214,72 +2215,4 @@ func labelNamesFromSeriesSet(series []*storepb.Series) []string {
 
 	sort.Strings(labels)
 	return labels
-}
-
-func TestInjectExtLabels(t *testing.T) {
-	testInjectExtLabels(testutil.NewTB(t))
-}
-
-func BenchmarkInjectExtLabels(b *testing.B) {
-	testInjectExtLabels(testutil.NewTB(b))
-}
-
-var x labels.Labels
-
-func testInjectExtLabels(tb testutil.TB) {
-	in := labels.FromStrings(
-		"__name__", "subscription_labels",
-		"_id", "0dfsdfsdsfdsffd1e96-4432-9abe-e33436ea969a",
-		"account", "1afsdfsddsfsdfsdfsdfsdfs",
-		"ebs_account", "1asdasdad45",
-		"email_domain", "asdasddgfkw.example.com",
-		"endpoint", "metrics",
-		"external_organization", "dfsdfsdf",
-		"instance", "10.128.4.231:8080",
-		"job", "sdd-acct-mngr-metrics",
-		"managed", "false",
-		"namespace", "production",
-		"organization", "dasdadasdasasdasaaFGDSG",
-		"pod", "sdd-acct-mngr-6669c947c8-xjx7f",
-		"prometheus", "telemeter-production/telemeter",
-		"prometheus_replica", "prometheus-telemeter-1",
-		"risk", "5",
-		"service", "sdd-acct-mngr-metrics",
-		"support", "Self-Support", // Should be overwritten.
-	)
-	extLset := labels.FromStrings(
-		"support", "Host-Support",
-		"replica", "1",
-		"tenant", "2342",
-	)
-	tb.ResetTimer()
-	for i := 0; i < tb.N(); i++ {
-		x = injectLabels(in, extLset)
-
-		if !tb.IsBenchmark() {
-			testutil.Equals(tb, labels.FromStrings(
-				"__name__", "subscription_labels",
-				"_id", "0dfsdfsdsfdsffd1e96-4432-9abe-e33436ea969a",
-				"account", "1afsdfsddsfsdfsdfsdfsdfs",
-				"ebs_account", "1asdasdad45",
-				"email_domain", "asdasddgfkw.example.com",
-				"endpoint", "metrics",
-				"external_organization", "dfsdfsdf",
-				"instance", "10.128.4.231:8080",
-				"job", "sdd-acct-mngr-metrics",
-				"managed", "false",
-				"namespace", "production",
-				"organization", "dasdadasdasasdasaaFGDSG",
-				"pod", "sdd-acct-mngr-6669c947c8-xjx7f",
-				"prometheus", "telemeter-production/telemeter",
-				"prometheus_replica", "prometheus-telemeter-1",
-				"replica", "1",
-				"risk", "5",
-				"service", "sdd-acct-mngr-metrics",
-				"support", "Host-Support",
-				"tenant", "2342",
-			), x)
-		}
-	}
-	fmt.Fprint(ioutil.Discard, x)
 }