--- a/pkg/store/labelpb/label.go
+++ b/pkg/store/labelpb/label.go
@@ -235,27 +235,35 @@ func (m *ZLabel) Compare(other ZLabel) int {
 	return strings.Compare(m.Value, other.Value)
 }
 
-// ExtendLabels extend given labels by extend in labels format.
+// ExtendSortedLabels extend given labels by extend in labels format.
 // The type conversion is done safely, which means we don't modify extend labels underlying array.
 //
 // In case of existing labels already present in given label set, it will be overwritten by external one.
-func ExtendLabels(lset labels.Labels, extend labels.Labels) labels.Labels {
-	overwritten := map[string]struct{}{}
-	for i, l := range lset {
-		if v := extend.Get(l.Name); v != "" {
-			lset[i].Value = v
-			overwritten[l.Name] = struct{}{}
+// NOTE: Labels and extend has to be sorted.
+func ExtendSortedLabels(lset labels.Labels, extend labels.Labels) labels.Labels {
+	ret := make(labels.Labels, 0, len(lset)+len(extend))
+
+	// Inject external labels in place.
+	for len(lset) > 0 && len(extend) > 0 {
+		d := strings.Compare(lset[0].Name, extend[0].Name)
+		if d == 0 {
+			// Duplicate, prefer external labels.
+			// NOTE(fabxc): Maybe move it to a prefixed version to still ensure uniqueness of series?
+			ret = append(ret, extend[0])
+			lset, extend = lset[1:], extend[1:]
+		} else if d < 0 {
+			ret = append(ret, lset[0])
+			lset = lset[1:]
+		} else if d > 0 {
+			ret = append(ret, extend[0])
+			extend = extend[1:]
 		}
 	}
 
-	for _, l := range extend {
-		if _, ok := overwritten[l.Name]; ok {
-			continue
-		}
-		lset = append(lset, l)
-	}
-	sort.Sort(lset)
-	return lset
+	// Append all remaining elements.
+	ret = append(ret, lset...)
+	ret = append(ret, extend...)
+	return ret
 }
 
 func PromLabelSetsToString(lsets []labels.Labels) string {